requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.8
lxml==4.5.0
Pillow==7.0.0
pydub==0.23.1
PyQt5==5.14.1
PyQt5-sip==12.7.1
pytesseract==0.3.2
python-docx==0.8.10
requests==2.22.0
six==1.14.0
soupsieve==1.9.5
SpeechRecognition==3.8.1
urllib3==1.25.8
